@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -41,7 +41,10 @@ dev_dependencies:
 # Switch to another stable flutter version
 dependency_overrides:
   fair_version:
-    path: ../../flutter_version/flutter_3_0_0
+    path: ../../flutter_version/flutter_3_10_0
+  analyzer: 5.6.0
+  dart_style: 2.2.5
+  collection: 1.17.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
